x-frame-options/commas.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>X-Frame-Options headers with commas</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.js"></script>

<body>
<script>
"use strict";

async_test(t => {
  const i = document.createElement("iframe");
  i.src = "./support/xfo.py?value=SAMEORIGIN,DENY";

  assert_no_message_from(i, t);

  i.onload = t.step_func_done(() => {
    assert_equals(i.contentDocument, null);
  });

  document.body.append(i);
  t.add_cleanup(() => i.remove());
}, "SAMEORIGIN,DENY blocks same-origin framing");

async_test(t => {
  const i = document.createElement("iframe");
  i.src = "./support/xfo.py?value=,SAMEORIGIN,,DENY,";

  assert_no_message_from(i, t);

  i.onload = t.step_func_done(() => {
    assert_equals(i.contentDocument, null);
  });

  document.body.append(i);
  t.add_cleanup(() => i.remove());
}, ",SAMEORIGIN,,DENY, blocks same-origin framing");

async_test(t => {
  const i = document.createElement("iframe");
  i.src = "./support/xfo.py?value=SAMEORIGIN,\"DENY\"";

  assert_no_message_from(i, t);

  i.onload = t.step_func_done(() => {
    assert_equals(i.contentDocument, null);
  });

  document.body.append(i);
  t.add_cleanup(() => i.remove());
}, "SAMEORIGIN,\"DENY\" blocks same-origin framing");

async_test(t => {
  const i = document.createElement("iframe");
  i.src = "./support/xfo.py?value=\"SAMEORIGIN,DENY\"";

  wait_for_message_from(i, t).then(t.step_func_done(e => {
    assert_equals(e.data, "Loaded");
  }));

  document.body.append(i);
  t.add_cleanup(() => i.remove());
}, "\"SAMEORIGIN,DENY\" allows same-origin framing");

async_test(t => {
  const i = document.createElement("iframe");
  i.src = "./support/xfo.py?value=  SAMEORIGIN,    DENY";

  assert_no_message_from(i, t);

  i.onload = t.step_func_done(() => {
    assert_equals(i.contentDocument, null);
  });

  document.body.append(i);
  t.add_cleanup(() => i.remove());
}, "  SAMEORIGIN,    DENY blocks same-origin framing");
</script>
